requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
beautifulsoup4
redis
aiohttp
aiofiles
aioredis
psutil
wikipedia
requests
Pillow
ujson
cleverwrap
xmltodict
youtube-dl